backend/voicebot-engine/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=14.0
httpx>=0.28.0
aiohttp>=3.9.0
google-generativeai>=0.3.0
